changed standardization, threshold, regularization
I have added a `logs/` directory to better track the progression of the
training process. Here are the exact changes (mentioned in the logs too)

- Added BatchNorm1d layer
- Set Dropout to 0.5
- add more metrics
- made the training focus on f-beta 0.5
- pos_iou_thresh=0.7,
- neg_iou_thresh=0.5,
- hard_neg_iou_range=(0.3, 0.5),
- num_neg_per_pos=5,
- change the standardization to be per channel

I have also added a [PROCESS.txt](PROCESS.txt) to give an overview of
the training process, which also helps with giving me clarity on what
exactly is the process and what can be optimized. This may be useful for
you to read if you want to understand the process too.